feat: terms api refactor use term.proto
tests moved to local tests, for work with contract

Closes #86
