+++ src/services/TermService.ts
@@ -1,13 +1,13 @@
 import termRepository from '../repositories/TermRepository';
-import { Term } from '../types';
+import { TermDBValue } from '../types';
 import { FacilitySubLevels } from './DBService';
 import facilityRepository from '../repositories/FacilityRepository';
 
 export class TermService {
-  public async getAllFacilityTerms(facilityId: string): Promise<Term[]> {
+  public async getAllFacilityTerms(facilityId: string): Promise<TermDBValue[]> {
     const ids = await termRepository.getAllTermIds(facilityId);
 
-    const terms = new Set<Term>();
+    const terms = new Set<TermDBValue>();
     for (const id of ids) {
       const term = await termRepository.getTerm(facilityId, id);
       if (term) {
@@ -21,7 +21,7 @@ export class TermService {
   public async setTerm(
     facilityId: string,
     termId: string,
-    term: Term
+    term: TermDBValue
   ): Promise<void> {
     await termRepository.setTerm(facilityId, termId, term);
     await termRepository.addTermToIndex(facilityId, termId);
@@ -36,13 +36,13 @@ export class TermService {
     facilityId: string,
     indexKey: FacilitySubLevels,
     itemId: string
-  ): Promise<Term[]> {
+  ): Promise<TermDBValue[]> {
     const ids = await termRepository.getAllItemTermIndexes(
       facilityId,
       indexKey,
       itemId
     );
-    const terms = new Set<Term>();
+    const terms = new Set<TermDBValue>();
     for (const id of ids) {
       const term = await termRepository.getTerm(facilityId, id);
       if (term) {
